refactor: headerloginbutton >> omnibartoolbar
